requirements.txt
openai-whisper
pyannote.audio
groq
gradio
soundfile
datasets
tqdm
python-dotenv
flask
faster-whisper
pyannote.audio
pydub

# git add .
# git commit -m "update"
# git push